requirements.txt
flask
paho-mqtt
apscheduler
pandas
